refactor: extract CommonsImageFetcher trait for reuse across test specs
