+++ test/services/LocalImageCacheServiceIntegrationSpec.scala
@@ -15,7 +15,7 @@ import java.nio.file.Files
 import scala.concurrent.duration._
 import scala.concurrent.{Await, ExecutionContext}
 
-class LocalImageCacheServiceIntegrationSpec extends Specification {
+class LocalImageCacheServiceIntegrationSpec extends Specification with CommonsImageFetcher {
 
   sequential
   skipAllUnless(sys.props.get("integration").contains("true"))
@@ -42,39 +42,6 @@ class LocalImageCacheServiceIntegrationSpec extends Specification {
       )
     )
     new LocalImageCacheService(config, system)
-  }
-
-  private val imageInfoProps = Set("size", "url", "mime")
-
-  /** Fetch up to `limit` eligible images from a Commons category via scalawiki. */
-  def fetchImagesFromCommons(category: String, limit: Int): Seq[Image] = {
-    val bot = MwBot.fromHost("commons.wikimedia.org")
-    val pages = Await.result(
-      bot
-        .page(category)
-        .imageInfoByGenerator(
-          "categorymembers",
-          "cm",
-          namespaces = Set(Namespace.FILE),
-          props      = imageInfoProps,
-          titlePrefix = None
-        )
-        .map(_.toSeq),
-      60.seconds
-    )
-    pages
-      .flatMap { page =>
-        page.images.headOption.flatMap { ii =>
-          for {
-            id  <- page.id
-            url <- ii.url
-            w   <- ii.width
-            h   <- ii.height
-          } yield Image(id, page.title, Some(url), ii.pageUrl, w, h, mime = ii.mime)
-        }
-      }
-      .filter(img => img.isImage && img.url.isDefined && img.width > 0 && img.height > 0)
-      .take(limit)
   }
 
   // ── tests ──────────────────────────────────────────────────────────────────
